feat: allow multi-line footnotes
The existing footnote-body-matching regex
requires a space after the label's colon;
i.e., the line needs to start like '[^label]: '.

This does not provide for multiline footnote bodies,
where the first line can end with the ':'
and is followed by one or more indented lines.

The regex pattern introduced in this commit
does not require the existence of subsequent lines,
nor that any subsequent lines be indented.

Date:      Fri Nov 24 15:11:48 2023 -0500

On branch heckman
Changes to be committed:
	modified:   CHANGELOG.md
	modified:   README.md
	modified:   package-lock.json
	modified:   package.json
	modified:   src/utils/index.ts
	modified:   test/workspace/test.md